requirements.txt
aniso8601==9.0.1
astroid==2.11.5
async-generator==1.10
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.2
Flask-HTTPAuth==4.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
outcome==1.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.13.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.5.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
